Compile LispLibrary from lisp/*.lisp
Remove the fixed LispLibrary variable from MalkuthOS.ino and add a
script which can be used to create the variable at compile time from the
content of the lisp/*.lisp files.

Before the first build and upon any subsequent changes to lisp/*.lisp
files, tools/compile-lisp.sh must be run. This tool will create a lisp.h
and lisp.cpp file with the content of lisp/*.lisp, so a recompile of the
C sources should then be done before re-upload.