Add pwd and cd commands but cd is kind of broken.
Lines passed to cd actually need to be parsed in a more involved way to
handle a .. appearing in the middle of the line. Right now it just
traverses upwards based on the number of "../" or such it sees, so
something like "cd ../otherdir" will just go up one directory and not
traverse into otherdir.